Admin point updates
Anyone can give points (but the points will be a randomly generated number between 0-10).
I can give any number of points with a specified argument after override-give. - It uses my ID to check if the user is allowed to give points.
I might create a list and add other users to it and the search if their ids are in the list in the future.